refactor(config): make config simplier
