+++ Cargo.toml
@@ -18,19 +18,9 @@ keywords = [
   "tool"
 ]
 
-[workspace]
-members = ["config"]
-resolver = "2"
-
-[workspace.package]
-version = "14.6.0"
-authors = [ "dark0dave" ]
-edition = "2024"
-
 [dependencies]
 anstyle = "1.0.10"
 clap = { version = "^4.5.49", features = ["derive", "env", "string"] }
-config = { path = "./config" }
 confy = "^2.0.0"
 env_logger = "^0.11.8"
 log = "^0.4.22"
