--- config/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "config"
-version.workspace = true
-authors.workspace = true
-edition.workspace = true
-
-[dependencies]
-anstyle = "1.0.10"
-clap = { version = "^4.5.49", features = ["derive", "env", "string"] }
-confy = "^2.0.0"
-env_logger = "^0.11.8"
-log = "^0.4.22"
-serde = { version = "^1.0.217", features = ["derive"] }
-serde_derive = "^1.0.152"
-
-[dev-dependencies]
-pretty_assertions = "^1.4.1"
-tempfile = "3"
