+++ src/config/weidu_log_options.rs
@@ -1,4 +1,4 @@
-use crate::log_options::LogOptions;
+use crate::config::log_options::LogOptions;
 
 pub struct WeiduLogOptions(Vec<LogOptions>);
