+++ src/main.rs
@@ -9,6 +9,9 @@ use installers::{eet_install, normal_install};
 use scan::components::scan_components;
 use scan::languages::scan_langauges;
 
+use utils::find_all_mods;
+
+mod config;
 mod installers;
 mod internal_log;
 mod parser;
@@ -28,8 +31,16 @@ fn main() -> ExitCode {
     log::debug!("{:?}", config.args);
 
     let status = match config.args.command {
-        CommandType::Normal(command) => normal_install(&command, config.parser.clone()),
-        CommandType::Eet(command) => eet_install(&command, config.parser.clone()),
+        CommandType::Normal(command) => normal_install(
+            &command,
+            config.parser.clone(),
+            &mut find_all_mods(&command.options.mod_directories, command.options.depth),
+        ),
+        CommandType::Eet(command) => eet_install(
+            &command,
+            config.parser.clone(),
+            &mut find_all_mods(&command.options.mod_directories, command.options.depth),
+        ),
         CommandType::Scan(Scan::Languages(command)) => scan_langauges(&command),
         CommandType::Scan(Scan::Components(command)) => scan_components(&command),
     };
