+++ src/weidu/install_block.rs
@@ -1,4 +1,4 @@
-use config::log_options::LogOptions;
+use crate::config::log_options::LogOptions;
 
 pub(crate) trait WeiduInstallBlock {
     fn generate_weidu_args(
